Начала исправлять ввод и начала выносить выпадающие списки в отдельный файл
